add powermock (with mockito), fix current (admin) unit tests namespace (and delete crashing scheduler tests)
