+++ alitheia/core/pom.xml
@@ -65,6 +65,10 @@
       </plugin>
     </plugins>
   </build>
+  
+  <properties>
+    <powermock.version>1.5.3</powermock.version>
+  </properties>
 
   <dependencies>
     <dependency>
@@ -157,6 +161,18 @@
       <artifactId>jettison</artifactId>
       <version>1.2</version>
       <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.powermock</groupId>
+      <artifactId>powermock-module-junit4</artifactId>
+      <version>${powermock.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.powermock</groupId>
+      <artifactId>powermock-api-mockito</artifactId>
+      <version>${powermock.version}</version>
+      <scope>test</scope>
     </dependency>
   </dependencies>
