--- alitheia/core/src/test/java/eu/sqooss/test/service/scheduler/TestJob.java
@@ -1,70 +1,0 @@
-/*
- * This file is part of the Alitheia system, developed by the SQO-OSS
- * consortium as part of the IST FP6 SQO-OSS project, number 033331.
- *
- * Copyright 2007 - 2010 - Organization for Free and Open Source Software,  
- *                Athens, Greece.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *
- *     * Redistributions in binary form must reproduce the above
- *       copyright notice, this list of conditions and the following
- *       disclaimer in the documentation and/or other materials provided
- *       with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-
-package eu.sqooss.test.service.scheduler;
-
-import eu.sqooss.service.scheduler.Job;
-
-/**
- * This a test job class.
- * It has the holy purpuse to print a string n times.
- *
- * @author Christoph Schleifenbaum
- */
-class TestJob extends Job
-{
-
-    private int n;
-    private String s;
-
-    /**
-     * Contructor creating a job printing string \a s \a n times.
-     */
-    public TestJob(int n, String s) {
-        this.n = n;
-        this.s = s;
-    }
-
-    public long priority() {
-        return 0;
-    }
-    
-    protected void run() throws Exception {
-        System.out.println("Testjob running!");
-        for (int i = 0; i < n; ++i) {   
-            Thread.sleep(500);
-            System.out.println(s);
-        }
-        System.out.println("Testjob finished!");
-    }
-}
